+++ src/sections/Main/Main.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import Map from './components/Map';
 import ToggleButtons from './components/ToggleButtons';
 import { Container } from './Main.styles';
-import Footer from '../Footer/footer/footer';
 
 function Main() {
   const [items, setItems] = useState([]);
@@ -32,7 +31,6 @@ function Main() {
           </option>
         ))}
       </select> */}
-      <Footer/>
     </Container>
   );
 }
